FIX: rename async.py -> _async.py to avoid syntax errors in 3.7
async has moved to a proper keyword resulting in syntax errors on
import.
